topics/gemma4/gemma4-dgx-devbox/requirements.txt
flyte[tui]>=2.0
flyteplugins-vllm
gradio>=5.0.0
openai>=1.50.0
python-dotenv
